backend/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.3.0
